+++ toolkit/tkit/memory/stack_allocator.cpp
@@ -7,8 +7,8 @@ namespace TKit
 StackAllocator::StackAllocator(const usize p_Size, const usize p_Alignment) noexcept
     : m_Size(p_Size), m_Remaining(p_Size)
 {
-    m_Buffer = static_cast<std::byte *>(AllocateAlignedPlatformSpecific(p_Size, p_Alignment));
-    m_Entries.reserve(p_Size / sizeof(Entry));
+    m_Buffer = static_cast<std::byte *>(AllocateAlignedPlatformSpecific(static_cast<size_t>(p_Size), p_Alignment));
+    m_Entries.reserve(p_Size / static_cast<usize>(sizeof(Entry)));
 }
 
 StackAllocator::~StackAllocator() noexcept
